requirements.txt
absl-py==0.7.0
astor==0.7.1
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
certifi==2018.11.29
chardet==3.0.4
cloudpickle==0.7.0
cycler==0.10.0
dask==1.1.1
decorator==4.3.2
-e git+https://github.com/marcoancona/DeepExplain.git@a4d6dcd849cd408d7f03615208ef94dab101cfeb#egg=deepexplain
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
idna==2.8
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==2.2.3
mock==2.0.0
networkx==2.2
numpy==1.16.1
opencv-contrib-python==4.0.0.21
pbr==5.1.2
Pillow==5.4.1
protobuf==3.6.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.1
PyYAML==3.13
requests==2.21.0
scikit-image==0.14.2
scipy==1.2.1
six==1.12.0
subprocess32==3.5.3
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
toolz==0.9.0
urllib3==1.24.1
Werkzeug==0.14.1
